Two bug fixes
Fix OSS-2608 and #OSS-1579.
The root cause of OSS-2608 is that the BeginPutObject will dispose the stream but the content might be uploading in the other thread. This could happen in any thread count (not necessarily 10 thread).
The root cause of OSS-1579 is to call IDirectory.Add on an existing key when the request is retried.