fix the crc bug in append object
The CRC returned from append object API is the whole object's CRC, while the CRC calculated in the original way is just for that uploaded part.

The fix is to allow the caller set the init CRC and the Crc64Stream will calculated the crc based on that. It's much faster than calling combine API.