two minor fixes for Unity
1) It turns out Unity (either under .net 3.5 or .net 4.6) has bug around chunked encoding---sometimes OSS server simply returns BAD request indicating it cannot understand the request.
And after CRC is enabled by default, the stream during upload will be inseekable and thus trigger the chunk encoding.
But in fact it's not necessary use chunk encoding in this case as we can still get the length correctly.
So this change is to change the condition of using chunk condition, by if it can get the length of stream and thus to avoid using chunk encoding in normal case.

2) Use the Mono approach under Unity inside AddInternal() of ServiceClientImpl.